+++ StabilityMatrix.Avalonia/ViewModels/PackageManager/PackageCardViewModel.cs
@@ -84,7 +84,10 @@ public partial class PackageCardViewModel : ProgressViewModel
         if (string.IsNullOrWhiteSpace(value?.PackageName))
             return;
 
-        if (value.PackageName == UnknownPackage.Key)
+        if (
+            value.PackageName == UnknownPackage.Key
+            || packageFactory.FindPackageByName(value.PackageName) is null
+        )
         {
             IsUnknownPackage = true;
             CardImageSource = "";
@@ -202,6 +205,10 @@ public partial class PackageCardViewModel : ProgressViewModel
 
                 EventManager.Instance.OnInstalledPackagesChanged();
             }
+
+            Text = "";
+            IsIndeterminate = false;
+            Value = 0;
         }
     }