+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning 2.0](https://semver.org/spec/v2
 ## v2.5.5
 ### Added
 - Added Spanish UI language options, thanks to Carlos Baena and Lautaroturina for the translations
+### Fixed
+- Fixed infinite progress wheel when package uninstall fails
 
 ## v2.5.4
 ### Fixed